+++ src/containers/Channel.tsx
@@ -435,6 +435,7 @@ class Channel extends React.Component<Props, State> {
             <Menu.Item onClick={this.togglePollEditor} icon="horizontal-bar-chart" text="Create Straw Poll" />
             {connected && <Menu.Item onClick={this.props.addMarker} icon="bookmark" text="Add marker" />}
             {connected && <Menu.Item onClick={this.toggleLogsExporter} icon="archive" text="Export logs" />}
+            {isMod && <Menu.Item onClick={this.openModView} icon="build" text="Mod view" />}
             {isBroadcaster && <Menu.Item onClick={this.openRewardsQueue} icon="stopwatch" text="Rewards queue" />}
             <ModerationMenuItems
               toggleFollowersOnly={this.toggleFollowersOnly}
@@ -762,8 +763,20 @@ class Channel extends React.Component<Props, State> {
    */
   private openRewardsQueue = () => {
     const { channel } = this.props
+
     if (!_.isNil(channel)) {
       Twitch.openRewardsQueue(channel)
+    }
+  }
+
+  /**
+   * Opens the Twitch mod view.
+   */
+  private openModView = () => {
+    const { channel } = this.props
+
+    if (!_.isNil(channel)) {
+      Twitch.openModView(channel)
     }
   }